处理mybatis-plus的升级兼容性

+++ guerlab-cloud-commons/src/main/java/net/guerlab/cloud/commons/api/Api.java
@@ -157,7 +157,7 @@ public interface Api<E, SP extends SearchParams> {
 	 */
 	@PostMapping(SELECT_COUNT_PATH)
 	@Operation(summary = "查询总记录数", security = @SecurityRequirement(name = Constants.TOKEN))
-	int selectCount(@Parameter(description = "搜索参数对象", required = true) @RequestBody SP searchParams);
+	long selectCount(@Parameter(description = "搜索参数对象", required = true) @RequestBody SP searchParams);
 
 	/**
 	 * 新增实体.
